requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.7
qrcode[pil]==7.4.2
openpyxl==3.1.2
python-dotenv==1.0.1
bcrypt==4.1.2
requests==2.31.0
passlib==1.7.4
pydantic-settings==2.1.0
user-agents==2.2.0
pillow>=10.2.0